# feat: update Lighthouse performance thresholds