Performance Thresholds Raised (doing it incrementally)

- Performance: 85% → 90% (blocks PR)
- LCP: 3000ms → 2500ms (blocks PR)
- CLS: 0.2 → 0.1 (blocks PR)
- FCP: 1500ms → 1200ms (new critical threshold)
- Accessibility: 90% → 95% (warning)
- Best Practices: 90% → 95% (warning)
- SEO: 90% → 95% (warning)

Current metrics:
- Performance: 100%
- LCP: 677ms (well under 2500ms limit)
- FCP: 537ms (well under 1200ms limit)
- CLS: 0.0 (perfect stability)

Based on the current metrics, we can raise the threshold even further, and I will do so next.